Potential fix for code scanning alert no. 6: Multiplication result converted to larger type
To fix the problem, we need to ensure that the multiplication is performed using the larger integer type (unsigned long) to avoid overflow. This can be achieved by casting one of the operands to unsigned long before performing the multiplication. Specifically, we will cast glyph_index to unsigned long in the expression glyph_index * entry_len.

+++ libfreetype2-2.10.1/src/cid/cidgload.c
@@ -110,7 +110,7 @@
 
 
       if ( FT_STREAM_SEEK( cid->data_offset + cid->cidmap_offset +
-                           glyph_index * entry_len )               ||
+                           (unsigned long)glyph_index * entry_len ) ||
            FT_FRAME_ENTER( 2 * entry_len )                         )
         goto Exit;
